@@ -1,25 +1,29 @@
-import React from "react";
-import logo from "./logo.svg";
+import { useState } from "react";
+
 import "./App.css";
 
 function App() {
+  const [count, setCount] = useState(0);
+
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
+    <>
+      <div></div>
+      <h1 className="bg-red-500">Vite + React</h1>
+      <div className="card">
+        <button
+          className="bg-green-700"
+          onClick={() => setCount((count) => count + 1)}
         >
-          Learn React
-        </a>
-      </header>
-    </div>
+          count is {count}
+        </button>
+        <p>
+          Edit <code>src/App.tsx</code> and save to test HMR
+        </p>
+      </div>
+      <p className="read-the-docs">
+        Click on the Vite and React logos to learn more
+      </p>
+    </>
   );
 }
 
